Fixed cache location
Until now, the cache file was created in the current working directory.
This however was undesired as the caching strategy would have worked as
intended only if the user would have initiated the application every
time from the same directory. Furthermore, a cache file would have been
created in each path where the app would have been initiated.

This commit replaces `process.cwd()` with `__dirname` so that the cache
file is created always in the cache module directory.