Bump minimum Node.js version to 10
This commit bumps the minimum supported Node.js version to `10.12.0`.
This version has been picked for its introduction of the `recursive`
option for `fs.mkDir`.

Additionally, the previous min version 8 was already erroneously
unsupported, in part because of some transient dependencies, in part
because of the `tsconfig.json` extending a config for version 10.

Also, to prevent using Node.js APIs that might not be available on all
supported versions, the `@types/node` package is downgraded to version
10. This way you will be warned when attempting to use an API that is
too recent and not always available.