Use proper platform-specific cache directory (#384)
This commit adds `getCacheDir`, a function that returns a
platform-specific, user-space cache directory and that is used when
creating an instance of `Cache`.

Not only this offers a consistent path for cache data, it also allows
operating systems to correctly clear cached data in accordance with
their default behavior or by a user's explicit request.

However, only the three major platforms are properly implemented. For
all other cases, a system temporary directory is used instead.