Streamline error surfacing and management
This commit refactors how errors are handled and communicated to the
user.

First, it makes the necessary changes so that all unrecoverable errors
eventually bubble up `src/index.ts`. It's easier to handle errors in one
place.

Second, the commit adds the `handleCommandError` and `ApplicationError`
utilities. `handleCommandError` is what is used in `src/index.ts` to
capture, interpret and display errors.