Display error spinner more consistently
Previously, the spinner would not be displayed at all if there is no API
key available, even if the application is in the data fetching phase.

This commit moves the spinner creation at the top of `callApi` so that
it can be set to display error if the API key is missing.