Improve error management (#396)
* Streamline error surfacing and management

This commit refactors how errors are handled and communicated to the
user.

First, it makes the necessary changes so that all unrecoverable errors
eventually bubble up `src/index.ts`. It's easier to handle errors in one
place.

Second, the commit adds the `handleCommandError` and `ApplicationError`
utilities. `handleCommandError` is what is used in `src/index.ts` to
capture, interpret and display errors.

* Use ApplicationError, add new error cases

* Use cfonts only after successful API calls

In case of errors, displaying a huge block of text is not useful. So,
this commit moves the various `cfont.say` statements after all required
API calls, so that fonts will be displayed only if the app is sure to
have all the data it needs.

* Display error spinner more consistently

Previously, the spinner would not be displayed at all if there is no API
key available, even if the application is in the data fetching phase.

This commit moves the spinner creation at the top of `callApi` so that
it can be set to display error if the API key is missing.

* Update text for NETWORK_UNREACHABLE

Co-authored-by: Alessandro Cifani <alessandro.cifani@gmail.com>

* Don't consider a cache error fatal

The application can work without caching, so there is no need to have an
`ErrorCode.CACHE_WRITE` case.

---------

Co-authored-by: Alessandro Cifani <alessandro.cifani@gmail.com>